fix build problem + prelim dockerfile
